Separate status and history from EDU program queue
