+++ Sts1CobcSw/EduProgramQueue.cpp
@@ -7,5 +7,4 @@ namespace sts1cobcsw
 {
 std::uint16_t queueIndex = 0U;
 etl::vector<EduQueueEntry, eduProgramQueueSize> eduProgramQueue{};
-RODOS::RingBuffer<StatusHistoryEntry, statusHistorySize> statusHistory{};
 }
